Consolidate cloud job queuing
