--- internal/saucecloud/worker.go
@@ -1,10 +1,0 @@
-package saucecloud
-
-import "github.com/saucelabs/saucectl/internal/job"
-
-type result struct {
-	suiteName string
-	browser   string
-	job       job.Job
-	err       error
-}
